Prevent bet amounts with decimals

+++ cogs/betting.py
@@ -129,7 +129,7 @@ class BetCommands(commands.Cog, name="Betting Commands"):
             bet_amount = self.get_balance(ctx.message.author)
         elif type(bet_amount) == str and "%" in bet_amount:
             perc = float(bet_amount.strip("%")) / 100
-            bet_amount = self.get_balance(ctx.message.author) * perc
+            bet_amount = int(self.get_balance(ctx.message.author) * perc)
 
         self.check_balance(ctx.message.author, bet_amount)  # Checks if initialized and has the proper balance
